Rename many classes to the Tarcha\WebKernel namespace
