+++ .gitignore
@@ -1,3 +1,4 @@
 tests/kernel/vendor
 tests/kernel/composer.lock
 vendor
+composer.lock
