+++ config/Common.php
@@ -1,5 +1,5 @@
 <?php
-namespace Aura\Web_Kernel\_Config;
+namespace Tarcha\WebKernel\_Config;
 
 use Aura\Di\Config;
 use Aura\Di\Container;
@@ -14,54 +14,18 @@ class Common extends Config
         $di->set('aura/web-kernel:router', $di->lazyNew('Aura\Router\Router'));
         $di->set('aura/web-kernel:dispatcher', $di->lazyNew('Aura\Dispatcher\Dispatcher'));
 
-        // Aura\Web\ResponseSender
-        $di->params['Aura\Web\ResponseSender'] = array(
-            'response' => $di->lazyGet('aura/web-kernel:response'),
-        );
-
-        // Aura\Web_Kernel\Abstract_Responder
-        $di->params['Aura\Web_Kernel\AbstractResponder'] = array(
-            'response' => $di->lazyGet('aura/web-kernel:response'),
-        );
-
-        // Aura\Web_Kernel\CaughtException
-        $di->params['Aura\Web_Kernel\CaughtException'] = array(
-            'request' => $di->lazyGet('aura/web-kernel:request'),
-            'responder' => $di->lazyNew('Aura\Web_Kernel\CaughtExceptionResponder'),
-        );
-
-        // Aura\Web_Kernel\MissingAction
-        $di->params['Aura\Web_Kernel\MissingAction'] = array(
-            'request' => $di->lazyGet('aura/web-kernel:request'),
-            'responder' => $di->lazyNew('Aura\Web_Kernel\MissingActionResponder'),
-        );
-
-        // Aura\Web_Kernel\MissingRoute
-        $di->params['Aura\Web_Kernel\MissingRoute'] = array(
-            'request' => $di->lazyGet('aura/web-kernel:request'),
-            'responder' => $di->lazyNew('Aura\Web_Kernel\MissingRouteResponder'),
-        );
-
-        // Aura\Web_Kernel\WebKernel
-        $di->params['Aura\Web_Kernel\WebKernel'] = array(
-            'router' => $di->lazyNew('Aura\Web_Kernel\WebKernelRouter'),
-            'dispatcher' => $di->lazyNew('Aura\Web_Kernel\WebKernelDispatcher'),
-            'response_sender' => $di->lazyNew('Aura\Web\ResponseSender'),
-        );
-
-        // Aura\Web_Kernel\WebKernelDispatcher
-        $di->params['Aura\Web_Kernel\WebKernelDispatcher'] = array(
-            'request' => $di->lazyGet('aura/web-kernel:request'),
-            'dispatcher' => $di->lazyGet('aura/web-kernel:dispatcher'),
-            'logger' => $di->lazyGet('aura/project-kernel:logger'),
-        );
+        $di->types['Aura\Web\Request'] = $di->lazyGet('aura/web-kernel:request');
+        $di->types['Aura\Web\Response'] = $di->lazyGet('aura/web-kernel:response');
+        $di->types['Aura\Router\Router'] = $di->lazyGet('aura/web-kernel:router');
+        $di->types['Aura\Dispatcher\Dispatcher'] = $di->lazyGet('aura/web-kernel:dispatcher');
 
         // Aura\Web_Kernel\WebKernelRouter
-        $di->params['Aura\Web_Kernel\WebKernelRouter'] = array(
-            'request' => $di->lazyGet('aura/web-kernel:request'),
-            'router' => $di->lazyGet('aura/web-kernel:router'),
-            'logger' => $di->lazyGet('aura/project-kernel:logger'),
-        );
+        $di->params['Tarcha\WebKernel\WebKernelRouter']['logger']
+            = $di->lazyGet('aura/project-kernel:logger');
+
+        // Aura\Web_Kernel\WebKernelDispatcher
+        $di->params['Tarcha\WebKernel\WebKernelDispatcher']['logger']
+            = $di->lazyGet('aura/project-kernel:logger');
 
     }
