+++ src/Responders/CaughtExceptionResponder.php
@@ -8,7 +8,7 @@
  * @license http://opensource.org/licenses/bsd-license.php BSD
  *
  */
-namespace Aura\Web_Kernel;
+namespace Tarcha\WebKernel\Responders;
 
 use Exception;
