+++ src/Responders/MissingActionResponder.php
@@ -8,7 +8,9 @@
  * @license http://opensource.org/licenses/bsd-license.php BSD
  *
  */
-namespace Aura\Web_Kernel;
+namespace Tarcha\WebKernel\Responders;
+
+use Tarcha\WebKernel\Responders\AbstractResponder;
 
 /**
  *
