+++ src/Responders/MissingRouteResponder.php
@@ -8,7 +8,9 @@
  * @license http://opensource.org/licenses/bsd-license.php BSD
  *
  */
-namespace Aura\Web_Kernel;
+namespace Tarcha\WebKernel\Resopnders;
+
+use Tarcha\WebKernel\Responders\AbstractResponder;
 
 /**
  *
