+++ src/WebKernel.php
@@ -8,9 +8,11 @@
  * @license http://opensource.org/licenses/bsd-license.php BSD
  *
  */
-namespace Aura\Web_Kernel;
+namespace Tarcha\WebKernel;
 
+use Tarcha\WebKernel\WebKernelRouter;
 use Aura\Web\ResponseSender;
+use Tarcha\WebKernel\WebKernelDispatcher;
 
 /**
  *
@@ -25,7 +27,7 @@ class WebKernel
      *
      * Web router logic.
      *
-     * @var WebKernelRouter
+     * @var Router
      *
      */
     protected $router;
@@ -34,7 +36,7 @@ class WebKernel
      *
      * Web dispatcher logic.
      *
-     * @var WebKernelDispatcher
+     * @var Dispatcher
      *
      */
     protected $dispatcher;
