+++ src/WebKernelDispatcher.php
@@ -8,7 +8,7 @@
  * @license http://opensource.org/licenses/bsd-license.php BSD
  *
  */
-namespace Aura\Web_Kernel;
+namespace Tarcha\WebKernel;
 
 use Aura\Web\Request;
 use Aura\Dispatcher\Dispatcher;
