+++ tests/kernel/bootstrap.php
@@ -1,7 +1,7 @@
 <?php
 error_reporting(E_ALL);
 
-$composer_autoload = __DIR__ . "/vendor/autoload.php";
+$composer_autoload = __DIR__ . "/../../vendor/autoload.php";
 if (! is_readable($composer_autoload)) {
     echo "Did not find 'vendor/autoload.php'." . PHP_EOL;
     echo "Try ./phpunit.sh instead of phpunit." . PHP_EOL;
@@ -9,7 +9,7 @@ if (! is_readable($composer_autoload)) {
 }
 
 require $composer_autoload;
-require dirname(dirname(__DIR__)) . '/autoload.php';
+//require dirname(dirname(__DIR__)) . '/autoload.php';
 
 // need the fake response sender from the Aura.Web tests
 require __DIR__ . "/vendor/aura/web/tests/unit/src/FakeResponseSender.php";
