+++ tests/kernel/composer.json
@@ -1,22 +1,10 @@
 {
-    "name": "aura/web-kernel",
+    "name": "web-kernel",
     "type": "library",
-    "description": "The kernel files for an Aura web project.",
-    "keywords": [
-        "web",
-        "kernel"
-    ],
-    "homepage": "https://github.com/auraphp/Aura.Web_Kernel",
-    "license": "BSD-2-Clause",
-    "authors": [
-        {
-            "name": "Aura.Web_Kernel Contributors",
-            "homepage": "https://github.com/auraphp/Aura.Web_Kernel/contributors"
-        }
-    ],
     "require": {
         "php": ">=5.4.0",
         "aura/project-kernel": "~2.0",
+        "aura/web-kernel": "~2.0",
         "aura/dispatcher": "~2.0",
         "aura/router": "~2.0",
         "aura/web": "~2.0"
@@ -24,19 +12,23 @@
     "autoload": {
         "psr-4": {
             "Aura\\Web_Kernel\\": "src/",
-            "Aura\\Web_Kernel\\_Config\\": "config/"
+            "Tarcha\\WebKernel\\_Config\\": "config/",
+            "Tarcha\\WebKernel\\": "src/"
         }
     },
     "extra": {
         "aura": {
             "type": "kernel",
             "config": {
-                "common": "Aura\\Web_Kernel\\_Config\\Common",
-                "web-kernel-test": "Aura\\Web_Kernel\\_Config\\WebKernelTest"
+                "common": "Tarcha\\WebKernel\\_Config\\Common",
+                "web-kernel-test2": "Tarcha\\WebKernel\\_Config\\WebKernelTest"
             }
         },
         "branch-alias": {
             "dev-develop-2": "2.0.x-dev"
         }
+    },
+    "scripts": {
+        "test": "cd tests/kernel; ./phpunit.sh"
     }
 }
